# Conflicts:
#	kafka-eagle-api/pom.xml
#	kafka-eagle-api/src/main/java/org/smartloli/kafka/eagle/api/email/html/HtmlMail.java
#	kafka-eagle-api/src/main/java/org/smartloli/kafka/eagle/api/im/IMService.java
#	kafka-eagle-api/src/main/java/org/smartloli/kafka/eagle/api/im/IMServiceImpl.java
#	kafka-eagle-common/pom.xml
#	kafka-eagle-common/src/main/java/org/smartloli/kafka/eagle/common/constant/JmxConstants.java
#	kafka-eagle-common/src/main/java/org/smartloli/kafka/eagle/common/protocol/BrokersInfo.java
#	kafka-eagle-common/src/main/java/org/smartloli/kafka/eagle/common/util/KConstants.java
#	kafka-eagle-common/src/main/resources/log4j.properties
#	kafka-eagle-common/src/main/resources/system-config.properties
#	kafka-eagle-core/pom.xml
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/KafkaService.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/KafkaServiceImpl.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/Mx4jService.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/Mx4jServiceImpl.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/ZkService.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/ZkServiceImpl.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/v2/BrokerService.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/factory/v2/BrokerServiceImpl.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/ipc/KafkaOffsetGetter.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/metrics/KafkaMetricsService.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/metrics/KafkaMetricsServiceImpl.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/sql/execute/KafkaConsumerAdapter.java
#	kafka-eagle-core/src/main/java/org/smartloli/kafka/eagle/core/sql/execute/KafkaSqlParser.java
#	kafka-eagle-core/src/test/java/org/smartloli/kafka/eagle/factory/TestKafkaServiceImpl.java
#	kafka-eagle-core/src/test/java/org/smartloli/kafka/eagle/ipc/TestKSql.java
#	kafka-eagle-core/src/test/java/org/smartloli/kafka/eagle/ipc/TestKafkaOffsetGetter.java
#	kafka-eagle-plugin/pom.xml
#	kafka-eagle-plugin/src/main/java/org/smartloli/kafka/eagle/plugin/font/KafkaEagleVersion.java
#	kafka-eagle-plugin/src/main/java/org/smartloli/kafka/eagle/plugin/util/JConstants.java
#	kafka-eagle-web/pom.xml
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/controller/OffsetController.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/controller/RoleController.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/controller/TopicController.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/dao/AlertDao.xml
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/dao/MBeanDao.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/dao/MBeanDao.xml
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/dao/UserDao.xml
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/quartz/AlertQuartz.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/quartz/MBeanQuartz.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/quartz/MetricsQuartz.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/MetricsService.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/OffsetService.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/TopicService.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/impl/MetricsServiceImpl.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/impl/OffsetServiceImpl.java
#	kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/impl/TopicServiceImpl.java
#	kafka-eagle-web/src/main/resources/bash/ke.sh
#	kafka-eagle-web/src/main/resources/conf/alert-quartz.xml
#	kafka-eagle-web/src/main/resources/conf/log4j.properties
#	kafka-eagle-web/src/main/resources/conf/system-config.properties
#	kafka-eagle-web/src/main/resources/kms/bin/catalina.sh
#	kafka-eagle-web/src/main/resources/kms/logs/Readme.md
#	kafka-eagle-web/src/main/resources/log4j.properties
#	kafka-eagle-web/src/main/resources/metrics-quartz.xml
#	kafka-eagle-web/src/main/resources/offsets-quartz.xml
#	kafka-eagle-web/src/main/resources/topic-quartz.xml
#	kafka-eagle-web/src/main/webapp/WEB-INF/views/consumers/offset_realtime.jsp
#	kafka-eagle-web/src/main/webapp/WEB-INF/views/metrics/kafka.jsp
#	kafka-eagle-web/src/main/webapp/WEB-INF/views/metrics/zk.jsp
#	kafka-eagle-web/src/main/webapp/WEB-INF/views/public/navbar.jsp
#	kafka-eagle-web/src/main/webapp/WEB-INF/views/system/user.jsp
#	kafka-eagle-web/src/main/webapp/WEB-INF/web.xml
#	kafka-eagle-web/src/main/webapp/media/js/main/consumer/offset.realtime.js
#	kafka-eagle-web/src/main/webapp/media/js/main/metrics/brokers.js
#	kafka-eagle-web/src/main/webapp/media/js/main/metrics/kafka.js
#	kafka-eagle-web/src/main/webapp/media/js/main/metrics/zk.js
#	kafka-eagle-web/src/main/webapp/media/js/main/topic/topic.meta.js
#	pom.xml

+++ kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/controller/TopicController.java
@@ -25,6 +25,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import org.apache.commons.lang.StringUtils;
+import org.apache.kafka.clients.admin.ConfigEntry;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
+import org.smartloli.kafka.eagle.common.protocol.MetadataInfo;
+import org.smartloli.kafka.eagle.common.protocol.PartitionsInfo;
+import org.smartloli.kafka.eagle.common.protocol.topic.TopicConfig;
+import org.smartloli.kafka.eagle.common.util.KConstants;
+import org.smartloli.kafka.eagle.common.util.KConstants.Kafka;
+import org.smartloli.kafka.eagle.common.util.KConstants.Topic;
+import org.smartloli.kafka.eagle.common.util.SystemConfigUtils;
+import org.smartloli.kafka.eagle.core.factory.KafkaFactory;
+import org.smartloli.kafka.eagle.core.factory.KafkaService;
+import org.smartloli.kafka.eagle.web.service.TopicService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -36,22 +49,6 @@ import org.springframework.web.servlet.ModelAndView;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
-
-import org.apache.kafka.clients.admin.ConfigEntry;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.smartloli.kafka.eagle.common.protocol.PartitionsInfo;
-import org.smartloli.kafka.eagle.common.protocol.topic.TopicConfig;
-import org.smartloli.kafka.eagle.common.util.KConstants;
-import org.smartloli.kafka.eagle.common.util.KConstants.Kafka;
-import org.smartloli.kafka.eagle.common.util.KConstants.Topic;
-import org.smartloli.kafka.eagle.common.util.SystemConfigUtils;
-import org.smartloli.kafka.eagle.core.factory.KafkaFactory;
-import org.smartloli.kafka.eagle.core.factory.KafkaService;
-import org.smartloli.kafka.eagle.core.factory.v2.BrokerFactory;
-import org.smartloli.kafka.eagle.core.factory.v2.BrokerService;
-import org.smartloli.kafka.eagle.core.metrics.KafkaMetricsFactory;
-import org.smartloli.kafka.eagle.core.metrics.KafkaMetricsService;
-import org.smartloli.kafka.eagle.web.service.TopicService;
 
 /**
  * Kafka topic controller to viewer data.
@@ -173,32 +170,58 @@ public class TopicController {
 		HttpSession session = request.getSession();
 		String clusterAlias = session.getAttribute(KConstants.SessionAlias.CLUSTER_ALIAS).toString();
 
-		String metadata = topicService.metadata(clusterAlias, tname);
-		JSONArray metadatas = JSON.parseArray(metadata);
-		int offset = 0;
+		Map<String, Object> map = new HashMap<>();
+		map.put("start", iDisplayStart);
+		map.put("length", iDisplayLength);
+		long count = topicService.getPartitionNumbers(clusterAlias, tname);
+		List<MetadataInfo> metadatas = topicService.metadata(clusterAlias, tname, map);
 		JSONArray aaDatas = new JSONArray();
-		for (Object object : metadatas) {
-			JSONObject meta = (JSONObject) object;
-			if (offset < (iDisplayLength + iDisplayStart) && offset >= iDisplayStart) {
-				JSONObject obj = new JSONObject();
-				obj.put("topic", tname);
-				obj.put("partition", meta.getInteger("partitionId"));
-				obj.put("logsize", meta.getInteger("logSize"));
-				obj.put("leader", meta.getInteger("leader"));
-				obj.put("replicas", meta.getString("replicas"));
-				obj.put("isr", meta.getString("isr"));
-				aaDatas.add(obj);
-			}
-			offset++;
+		for (MetadataInfo metadata : metadatas) {
+			JSONObject object = new JSONObject();
+			object.put("topic", tname);
+			object.put("partition", metadata.getPartitionId());
+			object.put("logsize", metadata.getLogSize());
+			object.put("leader", metadata.getLeader());
+			object.put("replicas", metadata.getReplicas());
+			object.put("isr", metadata.getIsr());
+			aaDatas.add(object);
 		}
 
 		JSONObject target = new JSONObject();
 		target.put("sEcho", sEcho);
-		target.put("iTotalRecords", metadatas.size());
-		target.put("iTotalDisplayRecords", metadatas.size());
+		target.put("iTotalRecords", count);
+		target.put("iTotalDisplayRecords", count);
 		target.put("aaData", aaDatas);
 		try {
 			byte[] output = target.toJSONString().getBytes();
+			BaseController.response(output, response);
+		} catch (Exception ex) {
+			ex.printStackTrace();
+		}
+	}
+
+	/** Get cluster data by ajax. */
+	@RequestMapping(value = "/topic/meta/mbean/{tname}/ajax", method = RequestMethod.GET)
+	public void topicMetaMetricsAjax(@PathVariable("tname") String tname, HttpServletResponse response, HttpServletRequest request, HttpSession session) {
+		try {
+			String clusterAlias = session.getAttribute(KConstants.SessionAlias.CLUSTER_ALIAS).toString();
+			String target = topicService.getTopicMBean(clusterAlias, tname);
+
+			byte[] output = target.getBytes();
+			BaseController.response(output, response);
+		} catch (Exception ex) {
+			ex.printStackTrace();
+		}
+	}
+
+	/** Get cluster data by ajax. */
+	@RequestMapping(value = "/topic/meta/jmx/{tname}/ajax", method = RequestMethod.GET)
+	public void topicMsgByJmxAjax(@PathVariable("tname") String tname, HttpServletResponse response, HttpServletRequest request, HttpSession session) {
+		try {
+			String clusterAlias = session.getAttribute(KConstants.SessionAlias.CLUSTER_ALIAS).toString();
+			String target = topicService.getTopicSizeAndCapacity(clusterAlias, tname);
+
+			byte[] output = target.getBytes();
 			BaseController.response(output, response);
 		} catch (Exception ex) {
 			ex.printStackTrace();
@@ -223,13 +246,13 @@ public class TopicController {
 
 	/** Get topic datasets by ajax. */
 	@RequestMapping(value = "/topic/manager/keys/ajax", method = RequestMethod.GET)
-	public void managerTopicKeysAjax(HttpServletResponse response, HttpServletRequest request) {
+	public void getTopicProperties(HttpServletResponse response, HttpServletRequest request) {
 		try {
 			HttpSession session = request.getSession();
 			String clusterAlias = session.getAttribute(KConstants.SessionAlias.CLUSTER_ALIAS).toString();
 			String name = request.getParameter("name");
 			JSONObject object = new JSONObject();
-			object.put("items", JSON.parseArray(topicService.listTopicKeys(clusterAlias, name)));
+			object.put("items", JSON.parseArray(topicService.getTopicProperties(clusterAlias, name)));
 			byte[] output = object.toJSONString().getBytes();
 			BaseController.response(output, response);
 		} catch (Exception ex) {
@@ -308,8 +331,13 @@ public class TopicController {
 		map.put("search", search);
 		map.put("start", iDisplayStart);
 		map.put("length", iDisplayLength);
-		long count = brokerService.topicNumbers(clusterAlias);
-		List<PartitionsInfo> topics = brokerService.topicRecords(clusterAlias, map);
+		long count = 0L;
+		if (search != null && !"".equals(search)) {
+			count = topicService.getTopicNumbers(clusterAlias, search);
+		} else {
+			count = topicService.getTopicNumbers(clusterAlias);
+		}
+		List<PartitionsInfo> topics = topicService.list(clusterAlias, map);
 		JSONArray aaDatas = new JSONArray();
 		for (PartitionsInfo partition : topics) {
 			JSONObject object = new JSONObject();
@@ -317,7 +345,6 @@ public class TopicController {
 			object.put("topic", "<a href='/ke/topic/meta/" + partition.getTopic() + "/' target='_blank'>" + partition.getTopic() + "</a>");
 			object.put("partitions", partition.getPartitions().size() > Topic.PARTITION_LENGTH ? partition.getPartitions().toString().substring(0, Topic.PARTITION_LENGTH) + "..." : partition.getPartitions().toString());
 			object.put("partitionNumbers", partition.getPartitionNumbers());
-			object.put("topicSize", "<a class='btn btn-primary btn-xs'>" + kafkaMetricsService.topicSize(clusterAlias, partition.getTopic()) + "</a>&nbsp");
 			object.put("created", partition.getCreated());
 			object.put("modify", partition.getModify());
 			if (Kafka.CONSUMER_OFFSET_TOPIC.equals(partition.getTopic())) {
@@ -366,7 +393,7 @@ public class TopicController {
 	@RequestMapping(value = "/topic/{topicName}/{token}/delete", method = RequestMethod.GET)
 	public ModelAndView topicDelete(@PathVariable("topicName") String topicName, @PathVariable("token") String token, HttpSession session, HttpServletResponse response, HttpServletRequest request) {
 		ModelAndView mav = new ModelAndView();
-		if (SystemConfigUtils.getProperty("kafka.eagle.topic.token").equals(token)) {
+		if (SystemConfigUtils.getProperty("kafka.eagle.topic.token").equals(token) && !Kafka.CONSUMER_OFFSET_TOPIC.equals(topicName)) {
 			String clusterAlias = session.getAttribute(KConstants.SessionAlias.CLUSTER_ALIAS).toString();
 			Map<String, Object> respons = kafkaService.delete(clusterAlias, topicName);
 			if ("success".equals(respons.get("status"))) {
@@ -441,6 +468,28 @@ public class TopicController {
 		target.put("aaData", aaDatas);
 		try {
 			byte[] output = target.toJSONString().getBytes();
+			BaseController.response(output, response);
+		} catch (Exception ex) {
+			ex.printStackTrace();
+		}
+	}
+
+	/** Get producer chart data by ajax. */
+	@RequestMapping(value = "/topic/producer/chart/ajax", method = RequestMethod.GET)
+	public void topicProducerChartAjax(HttpServletResponse response, HttpServletRequest request, HttpSession session) {
+		try {
+			String clusterAlias = session.getAttribute(KConstants.SessionAlias.CLUSTER_ALIAS).toString();
+
+			Map<String, Object> param = new HashMap<>();
+			param.put("cluster", clusterAlias);
+			param.put("stime", request.getParameter("stime"));
+			param.put("etime", request.getParameter("etime"));
+			param.put("topic", request.getParameter("topic"));
+			String target = topicService.queryTopicProducerChart(param);
+			if (StringUtils.isEmpty(target)) {
+				target = "";
+			}
+			byte[] output = target.getBytes();
 			BaseController.response(output, response);
 		} catch (Exception ex) {
 			ex.printStackTrace();
